requirements.txt
fastapi[standard]
uvicorn
python-dotenv
pydantic
langchain
langchain-openai
langchain-community
langchain-text-splitters
pypdf
httpx
psycopg[binary] 
